requirements.txt
python-dotenv
pymupdf
langchain-core
langchain-openai
langchain
langchain-teddynote
langchain-community
chromadb
langgraph
yaml